Use sized integer types for game logic

+++ src/game.h
@@ -1,3 +1,5 @@
-int game_loop();
-int init_game();
-int terminate_game();
+#include "definitions.h"
+
+int8_t game_loop();
+int8_t init_game();
+int8_t terminate_game();
